agri-azure-simulation/requirements.txt
azure-core==1.33.0
azure-iot-device==2.14.0
azure-iot-hub==2.6.1
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
deprecation==2.1.0
fonttools==4.57.0
idna==3.10
isodate==0.7.2
janus==2.0.0
kiwisolver==1.4.8
matplotlib==3.10.1
msrest==0.7.1
numpy==2.2.4
oauthlib==3.2.2
packaging==24.2
paho-mqtt==1.6.1
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-unixsocket2==1.0.0
six==1.17.0
typing_extensions==4.13.2
tzdata==2025.2
uamqp==1.6.11
urllib3==2.4.0
